Avoid re render of table row on hover
New method keep the index of row but it does not render, since hoverIndex is inside an object it will not trigger re render until a new object is set.

onMouseEnter -> It will set the hoverIndex to null so previous index won't show, and there won't be two active row.

onMouseLeave -> It set hoverIndex and force to re render, so when mouse hover is outside the table body it show the row that last of hovered.

onhover in css will take care of when mouse is on table row.